Moved data to Stories
